# Improve scan kernel