@anaruse 2020-11-25
Attaching the figure which shows the computation network used in `inclusive_scan_kernel_shfl`.

[shfl-scan-kernel.pdf](https://github.com/cupy/cupy/files/5594935/shfl-scan-kernel.pdf)

Hope this may help your understanding.