@anaruse 2020-11-26
Corrected the figure. This might be easier to understand.

[shfl-scan-kernel-2.pdf](https://github.com/cupy/cupy/files/5600027/shfl-scan-kernel-2.pdf)