@anaruse 2020-11-27
I think it's better to review the current inclusive scan implementation first, so I've made this PR back to draft.